made a minor modification in insurance controller

+++ back-end/src/main/java/com/example/backend/controller/DTO/ClientDTO.java
@@ -17,7 +17,7 @@ public class ClientDTO {
     private int clientId=0;
 
     @ApiModelProperty(required = true)
-    private short status = 2;
+    private int status = 2;
 
     @ApiModelProperty
     private String firstname;
